style(typings): import FlextTypes at runtime in flext-api/base_service
Move FlextTypes import out of TYPE_CHECKING to top-level.

+++ src/flext_api/api_exceptions.py
@@ -394,7 +394,7 @@ class FlextApiTimeoutError(FlextTimeoutError):
             context["operation"] = operation
         context["status_code"] = status_code
 
-        # Include timeout_seconds in context for test expectations
+        # Include timeout_seconds in context for test expectations with full precision
         if timeout_seconds is not None:
             context["timeout_seconds"] = float(timeout_seconds)
         # Avoid passing duplicate 'timeout_seconds' when also included in **context to base class
@@ -404,7 +404,7 @@ class FlextApiTimeoutError(FlextTimeoutError):
             message=message,
             service="flext_api_service",
             timeout_seconds=(
-                int(timeout_seconds) if timeout_seconds is not None else None
+                float(timeout_seconds) if timeout_seconds is not None else None
             ),
             context=init_kwargs,
         )
